Moved epoch class into its own file, adhering to internal coding standards.
