+++ NEventstore.Persistence.DocumentDB/DocumentCommit.cs
@@ -33,18 +33,4 @@
 
         public long CheckpointNumber { get; set; }
     }
-
-    internal class DateEpoch
-    {
-        public DateTime Date { get; set; }
-        public long Epoch
-        {
-            get
-            {
-                return (this.Date.Equals(null) || this.Date.Equals(DateTime.MinValue))
-                    ? int.MinValue
-                    : this.Date.ToEpoch();
-            }
-        }
-    }
 }
