+++ NEventstore.Persistence.DocumentDB/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.Azure.DocumentDB" version="1.7.1" targetFramework="net452" />
-  <package id="NEventStore" version="5.2.0" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net452" />
+  <package id="Microsoft.Azure.DocumentDB" version="1.7.1" targetFramework="net451" />
+  <package id="NEventStore" version="5.2.0" targetFramework="net451" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net451" />
 </packages>
